Adaptive-TCP-algorithm-based-on-machine-learning-in-telecommunication-networks
On the Internet, many devices use the TCP protocol for reliable data transfer. The main problems with network transmission are window size overload and packet loss. In TCP the sender considers network loss as the loss of packets in a transmission queue caused by network congestion. The results in poor efficiency and throughput. TCP congestion control is the problem of determining the correct size of the congestion window, which is the number of packets that the sender can send without receiving confirmation of the delivered packet. On the one hand, if the window size is too small, then the sender will not make full use of the available bandwidth. On the other hand, if the window size is too large, the packets will queue at the routers, which will increase the transit time of each packet and possibly for packet loss.